Change Steps structure
