+++ src/components/pages/manage/movies/create.tsx
@@ -1,67 +1,127 @@
 import { useMachine } from '@xstate/react';
-import type { MovieContext, MovieEvent } from 'machines/movie';
-import { movieMachine } from 'machines/movie';
+import { useMemo } from 'react';
 import { useTranslation } from 'react-i18next';
 import type { State } from 'xstate';
+import { assign, createMachine } from 'xstate';
 
+import type { MovieThemes } from '@/constants/MovieThemes';
+import { OfferCategories } from '@/constants/OfferCategories';
+import { OfferType } from '@/constants/OfferType';
 import { useLog } from '@/hooks/useLog';
-import { Box } from '@/ui/Box';
+import type { Place } from '@/types/Place';
+import type { Production } from '@/types/Production';
+import type { Values } from '@/types/Values';
 import { Page } from '@/ui/Page';
-import type { StackProps } from '@/ui/Stack';
-import { getStackProps, Stack } from '@/ui/Stack';
-import { Text } from '@/ui/Text';
-import { getValueFromTheme } from '@/ui/theme';
-import { Title } from '@/ui/Title';
 import { getApplicationServerSideProps } from '@/utils/getApplicationServerSideProps';
 
-import { Step1Content } from './Step1Content';
-import { Step2Content } from './Step2Content';
-import { Step3Content } from './Step3Content';
-import { Step4Content } from './Step4Content';
+import { Step1 } from './Step1';
+import { Step2 } from './Step2';
+import { Step3 } from './Step3';
+import { Step4 } from './Step4';
 
-const getValue = getValueFromTheme('moviesCreatePage');
+const MovieEventTypes = {
+  CHOOSE_THEME: 'CHOOSE_THEME',
+  CLEAR_THEME: 'CLEAR_THEME',
+  CHANGE_TIME_TABLE: 'CHANGE_TIME_TABLE',
+  CHOOSE_CINEMA: 'CHOOSE_CINEMA',
+  CLEAR_CINEMA: 'CLEAR_CINEMA',
+  CHOOSE_PRODUCTION: 'CHOOSE_PRODUCTION',
+  CLEAR_PRODUCTION: 'CLEAR_PRODUCTION',
+} as const;
 
-type StepProps = StackProps & { step: number };
+type Theme = Values<typeof MovieThemes>;
+type Time = string;
+
+type MovieContext = {
+  offerType: typeof OfferType.EVENT;
+  type: typeof OfferCategories.Film;
+  theme: Theme;
+  timeTable: Time[][];
+  cinema: Place;
+  production: Production;
+};
+
+type MovieEvent =
+  | { type: typeof MovieEventTypes.CHOOSE_THEME; value: Theme }
+  | { type: typeof MovieEventTypes.CLEAR_THEME }
+  | { type: typeof MovieEventTypes.CHANGE_TIME_TABLE; value: Time[][] }
+  | { type: typeof MovieEventTypes.CHOOSE_CINEMA; value: Place }
+  | { type: typeof MovieEventTypes.CLEAR_CINEMA }
+  | { type: typeof MovieEventTypes.CHOOSE_PRODUCTION; value: Production }
+  | { type: typeof MovieEventTypes.CLEAR_PRODUCTION };
 
 type MachineProps = {
   movieState: State<MovieContext, MovieEvent>;
   sendMovieEvent: (event: MovieEvent) => State<MovieContext, MovieEvent>;
 };
 
-const Step = ({ step, children, ...props }: StepProps) => {
-  const { t } = useTranslation();
-
-  return (
-    <Stack spacing={4} {...getStackProps(props)}>
-      <Title
-        color={getValue('title.color')}
-        lineHeight="220%"
-        alignItems="center"
-        spacing={3}
-        css={`
-          border-bottom: 1px solid ${getValue('title.borderColor')};
-        `}
-      >
-        <Box
-          borderRadius="50%"
-          width="1.8rem"
-          height="1.8rem"
-          lineHeight="1.8rem"
-          backgroundColor={getValue('stepNumber.backgroundColor')}
-          padding={0}
-          fontSize="1rem"
-          fontWeight="bold"
-          color="white"
-          textAlign="center"
-        >
-          {step}
-        </Box>
-        <Text>{t(`movies.create.step${step}_title`)}</Text>
-      </Title>
-      {children}
-    </Stack>
-  );
-};
+const movieMachine = createMachine<MovieContext, MovieEvent>({
+  id: 'movie',
+  initial: 'idle',
+  context: {
+    offerType: OfferType.EVENT,
+    type: OfferCategories.Film,
+    theme: null,
+    timeTable: null,
+    cinema: null,
+    production: null,
+  },
+  states: {
+    idle: {
+      on: {
+        [MovieEventTypes.CHOOSE_THEME]: {
+          actions: [
+            assign({
+              theme: (ctx, event) => event.value,
+            }),
+          ],
+        },
+        [MovieEventTypes.CLEAR_THEME]: {
+          actions: [
+            assign({
+              theme: () => null,
+            }),
+          ],
+        },
+        [MovieEventTypes.CHANGE_TIME_TABLE]: {
+          actions: [
+            assign({
+              timeTable: (ctx, event) => event.value,
+            }),
+          ],
+        },
+        [MovieEventTypes.CHOOSE_CINEMA]: {
+          actions: [
+            assign({
+              cinema: (ctx, event) => event.value,
+            }),
+          ],
+        },
+        [MovieEventTypes.CLEAR_CINEMA]: {
+          actions: [
+            assign({
+              cinema: () => null,
+            }),
+          ],
+        },
+        [MovieEventTypes.CHOOSE_PRODUCTION]: {
+          actions: [
+            assign({
+              production: (ctx, event) => event.value,
+            }),
+          ],
+        },
+        [MovieEventTypes.CLEAR_PRODUCTION]: {
+          actions: [
+            assign({
+              production: () => null,
+            }),
+          ],
+        },
+      },
+    },
+  },
+});
 
 const Create = () => {
   const [movieState, sendMovieEvent] = useMachine(movieMachine);
@@ -70,22 +130,22 @@ const Create = () => {
 
   useLog({ production: movieState.context.production });
 
+  const steps = useMemo(() => [Step1, Step2, Step3, Step4], []);
+
   return (
     <Page>
       <Page.Title spacing={3} alignItems="center">
         {t(`movies.create.title`)}
       </Page.Title>
       <Page.Content spacing={5}>
-        {[Step1Content, Step2Content, Step3Content, Step4Content].map(
-          (StepContent, index) => (
-            <Step key={index} step={index + 1}>
-              <StepContent
-                movieState={movieState}
-                sendMovieEvent={sendMovieEvent}
-              />
-            </Step>
-          ),
-        )}
+        {steps.map((Step, index) => (
+          <Step
+            movieState={movieState}
+            sendMovieEvent={sendMovieEvent}
+            stepNumber={index + 1}
+            key={index}
+          />
+        ))}
       </Page.Content>
     </Page>
   );
@@ -94,4 +154,5 @@ const Create = () => {
 export const getServerSideProps = getApplicationServerSideProps();
 
 export type { MachineProps };
+export { MovieEventTypes };
 export default Create;
